requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bcrypt==3.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.8.3
networkx==3.3
nltk==3.9.1
numpy==2.1.1
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
protobuf==5.28.2
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.6
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
python-jose
azure.storage.blob